requirements.txt
fastapi
requests
uvicorn
httpx
python-dotenv
openai
pydantic